backend/requirements.txt
py-bcrypt
pydantic[email]

fastapi
uvicorn[standard]
pydantic[email]
pymongo
python-dotenv
passlib[bcrypt]
python-jose[cryptography]



